Move all commands to cmds submodule, remove some deprecated commands
