+++ cheeto/config.py
@@ -7,20 +7,14 @@
 # Author : Camille Scott <cswel@ucdavis.edu>
 # Date   : 11.10.2024
 
-from argparse import Namespace
-import atexit
 import logging
-import os
 import pathlib
-import sys
 from typing import List, Optional, Union, Mapping
 
 from marshmallow.exceptions import ValidationError
 from marshmallow_dataclass import dataclass
 
-from . import log, __version__
-from .args import commands, ArgParser
-from .errors import ExitCode
+from . import __version__
 from .yaml import parse_yaml
 from .types import *
 from .utils import require_kwargs
@@ -123,84 +117,3 @@ def get_config(config_path: Optional[pathlib.Path] = None,
             hippo = config.hippo,
             ucdiam = config.ucdiam
         )
-
-
-@commands.root.args('common config', common=True)
-def common_args(parser: ArgParser):
-    parser.add_argument('--log',
-                       type=Path,
-                       default=Path(os.devnull),
-                       help='Log to file.')
-    parser.add_argument('--quiet',
-                       default=False,
-                       action='store_true')
-    parser.add_argument('--config',
-                       type=Path,
-                       default=DEFAULT_CONFIG_PATH,
-                       help='Path to alternate config file')
-    parser.add_argument('--profile',
-                       default='default',
-                       help='Config profile to use')
-
-
-@common_args.postprocessor(priority=0)
-def print_version(args: Namespace):
-    console = log.Console(stderr=True)
-    if not args.quiet:
-        console.print(f'cheeto [green]v{__version__}[/green]')
-
-
-@common_args.postprocessor(priority=100)
-def parse_config(args: Namespace):
-    args.config = get_config(config_path=args.config, profile=args.profile)
-    if 'accounts.hpc' in args.config.mongo.uri:
-        #pass
-        print("Testing right now, don't use prod", file=sys.stderr)
-        sys.exit(1)
-
-
-@common_args.postprocessor(priority=200)
-def setup_log(args: Namespace):
-    if args.log:
-        log_file = args.log.open('a')
-        log.setup(log_file, quiet=args.quiet)
-        
-        def close():
-            if log_file and not log_file.closed:
-                log_file.close()
-        atexit.register(close)
-
-
-@commands.register('config', 'show',
-                   help='Parse and show the config file')
-def show(args: Namespace):
-    logger = logging.getLogger(__name__)
-
-    if args.config is None:
-        sys.exit(ExitCode.VALIDATION_ERROR)
-    else:
-        print(Config.Schema().dumps(args.config))
-
-
-@commands.register('config', 'write',
-                   help='Write a skeleton config file')
-def write(args: Namespace):
-    logger = logging.getLogger(__name__)
-
-    config = Config(ldap = dict(
-                        hpccf = LDAPConfig(servers=['ldaps://ldap1.hpc.ucdavis.edu', 'ldaps://ldap2.hpc.ucdavis.edu'],
-                                           searchbase='dc=hpc,dc=ucdavis,dc=edu',
-                                           login_dn='uid=cheeto,ou=Services,dc=hpc,dc=ucdavis,dc=edu',
-                                           password='password',
-                                           user_classes=['inetOrgPerson', 'posixAccount']),
-
-                    ))
-
-    config_path = get_config_path()
-    if not config_path.exists():
-        logger.info(f'Config file not found, writing basic config to {config_path}')
-        config_path.parent.mkdir(parents=True, exist_ok=True)
-        config_path.write_text(Config.Schema().dumps(config))
-    else:
-        logger.warn(f'Config file already exists at {config_path}, exiting.')
-        sys.exit(0)
