+++ cheeto/monitor.py
@@ -7,13 +7,10 @@
 # Author : Camille Scott <cswel@ucdavis.edu>
 # Date   : 01.05.2023
 
-from argparse import Namespace
 from datetime import datetime
 import time
 
 import sh
-
-from .args import commands, ArgParser
 
 
 def parse_dcmi_power(dcmi_str: str):
@@ -22,23 +19,15 @@ def parse_dcmi_power(dcmi_str: str):
     return int(reading.removesuffix('Watts').strip())
 
 
-@commands.register('monitor', 'power',
-                   help='Monitor power usage from DCMI')
-def power(args: Namespace):
+def poll_dcmi_power(output: str, interval: int):
     cmd = sh.sudo.bake('ipmitool', 'dcmi', 'power', 'reading')
     
-    with open(args.output, 'w') as fp:
+    with open(output, 'w') as fp:
         try:
             while(True):
                 cur_time = datetime.now().isoformat()
                 cur_power = parse_dcmi_power(cmd())
                 print(f'{cur_time}, {cur_power}', file=fp, flush=True)
-                time.sleep(args.interval)
+                time.sleep(interval)
         except KeyboardInterrupt:
             pass
-
-
-@power.args()
-def add_power_args(parser: ArgParser):
-    parser.add_argument('-o', '--output', default='/dev/stdout')
-    parser.add_argument('--interval', default=5, type=int)
