+++ cheeto/puppet.py
@@ -7,26 +7,18 @@
 # Author : Camille Scott <cswel@ucdavis.edu>
 # Date   : 21.02.2023
 
-import argparse
 from dataclasses import field
 from enum import Enum
 import logging
-import os
-import traceback
 from typing import Callable, Optional, List, Mapping, Union, Tuple, Type
 from typing_extensions import Concatenate
 import sys
 
 from filelock import FileLock
-import ldap
 import marshmallow
 from marshmallow import pre_dump
 from marshmallow_dataclass import dataclass
-from rich import print as rprint
-from rich.console import Console
-from rich.syntax import Syntax
 
-from .args import commands, ArgParser, arggroup, EnumAction
 from .errors import ExitCode
 from .ldap import LDAPManager
 from .yaml import (MergeStrategy,
@@ -608,257 +600,5 @@ class SiteData(YamlRepo):
             yield group_name, group_record
 
 
-@arggroup('YAML Validation')
-def validate_args(parser: ArgParser):
-    parser.add_argument('--dump', default='/dev/stdout',
-                        help='Dump the validated YAML to the given file')
-    parser.add_argument('--echo', action='store_true', default=False)
-    parser.add_argument('files', nargs='+',
-                        help='YAML files to validate.')
-    parser.add_argument('--merge',
-                        default=MergeStrategy.NONE,
-                        type=MergeStrategy,
-                        action=EnumAction,
-                        help='Merge the given YAML files before validation.')
-    parser.add_argument('--strict', action='store_true', default=False,
-                        help='Terminate on validation errors.')
-    parser.add_argument('--partial',
-                        default=False,
-                        action='store_true',
-                        help='Allow partial loading (ie missing keys).')
-    parser.add_argument('--postload-validate', default=False,
-                       action='store_true')
-    for func_name in _postload_validators.keys():
-        parser.add_argument(f'--{func_name}', action='store_true', default=False)
-
-
-
-@validate_args.apply()
-@commands.register('puppet', 'validate',
-                   help='Validate a puppet.hpc-formatted YAML file')
-def validate_yamls(args: argparse.Namespace):
-
-    console = Console(stderr=True)
-
-    yaml_forest = parse_yaml_forest(args.files,
-                                    merge_on=args.merge)
-    
-    for source_file, puppet_data in validate_yaml_forest(yaml_forest,
-                                                         PuppetAccountMap,
-                                                         args.strict):
-        if not args.quiet: 
-            console.rule(source_file, style='blue')
-
-        if args.postload_validate:
-            for postload_func in _postload_validators.values():
-                postload_func(source_file, puppet_data, args.strict)
-        else:
-            for func_name in _postload_validators.keys():
-                if getattr(args, func_name):
-                    _postload_validators[func_name](source_file, puppet_data, args.strict)
-
-        output_yaml = PuppetAccountMap.Schema().dumps(puppet_data) 
-        hl_yaml = Syntax(output_yaml,
-                         'yaml',
-                         theme='github-dark',
-                         background_color='default')
-
-        if args.dump:
-            with open(args.dump, 'w') as fp:
-                if args.dump == '/dev/stdout':
-                    rprint(hl_yaml, file=fp)
-                else:
-                    print(output_yaml, file=fp)
-
-        if args.dump != '/dev/stdout' and args.echo and not args.quiet:
-            console.print(hl_yaml)
-
-
-class LDAPQueryParams(Enum):
-    username = 'uid'
-    uuid = 'ucdPersonUUID'
-    email = 'mail'
-    displayname = 'displayName'
-
-
 def flatten_and_decode(data: dict) -> dict:
     return {key: value[0].decode() for key, value in data.items()} #type: ignore
-
-
-@commands.register('puppet', 'create-nologin-user', 
-                    help='Create a nologin user')
-def create_nologin_user(args: argparse.Namespace):
-    console = Console(stderr=True)
-    logger = logging.getLogger(__name__)
-    conn = ldap.initialize(args.ldap_uri)
-    
-    query = ','.join((f'{param.value}={getattr(args, param.name)}' for param in LDAPQueryParams \
-                      if getattr(args, param.name) is not None))
-
-    console.print(f'Server: {args.ldap_uri}')
-    console.print(f'Query: {query}')
-
-    try:
-        result = conn.search_s('ou=People,dc=ucdavis,dc=edu',
-                               ldap.SCOPE_SUBTREE, #type: ignore
-                               query,
-                               [param.value for param in LDAPQueryParams])
-    except Exception as e:
-        console.print(e)
-    else:
-        if not (0 < len(result) < 2): #type: ignore
-            logger.error(f'Should get exactly one result (got {len(result)})') #type: ignore
-            logger.error(result)
-            sys.exit(ExitCode.BAD_LDAP_QUERY)
-        
-        data = flatten_and_decode(result[0][1]) #type: ignore
-        console.print(f'Result: {data}')
-
-        user_name = data['uid']
-        yaml_dumper = PuppetUserMap.global_dumper()
-        yaml_filename = args.global_dir / f'{user_name}.yaml'
-
-        if not args.force and os.path.exists(yaml_filename):
-            logger.error(f'{yaml_filename} already exists! Exiting.')
-            sys.exit(ExitCode.FILE_EXISTS)
-
-        user_record = PuppetUserMap(
-            user = {
-                user_name: PuppetUserRecord(
-                    fullname = data['displayName'],
-                    email = data.get('mail', ['donotreply@ucdavis.edu']),
-                    uid = data['ucdPersonUUID'],
-                    gid = data['ucdPersonUUID'],
-                )
-            }
-        )
-
-        output_yaml = yaml_dumper.dumps(user_record) #type: ignore
-        hl_yaml = Syntax(output_yaml,
-                         'yaml',
-                         theme='github-dark',
-                         background_color='default')
-
-        if not args.quiet:
-            console.print(hl_yaml)
-
-        with yaml_filename.open('w') as fp:
-            print(output_yaml, file=fp)
-
-        try:
-            link_relative(args.site_dir, yaml_filename)
-        except FileExistsError:
-            logger.info(f'{yaml_filename} already linked to {args.site_dir}.')
-
-
-@create_nologin_user.args()
-def add_create_nologin_user_args(parser: ArgParser):
-    qgroup = parser.add_argument_group('LDAP Query Parameters')
-    xgroup = qgroup.add_mutually_exclusive_group(required=True)
-    for param in LDAPQueryParams:
-        xgroup.add_argument(f'--{param.name}', metavar=param.value)
-    
-    lgroup = parser.add_argument_group('LDAP Server')
-    lgroup.add_argument('--ldap-uri', default='ldap://ldap.ucdavis.edu')
-
-    parser.add_argument('--site-dir', 
-                        type=Path,
-                        required=True,
-                        help='Site-specific puppet accounts directory.')
-    parser.add_argument('--global-dir',
-                        type=Path,
-                        required=True,
-                        help='Global puppet accounts directory.')
-    parser.add_argument('--force',
-                        default=False,
-                        action='store_true',
-                        help='Overwrite existing global YAML file.')
-
-
-@arggroup('YAML Repo Args')
-def repo_args(parser: ArgParser):
-    parser.add_argument('--site-dir', 
-                        type=Path,
-                        required=True,
-                        help='Site-specific puppet accounts directory.')
-    parser.add_argument('--global-dir',
-                        type=Path,
-                        required=True,
-                        help='Global puppet accounts directory.')
-    parser.add_argument('--key-dir',
-                        type=Path,
-                        required=True,
-                        help='Puppet SSH keys directory.')
-    parser.add_argument('--base-branch',
-                        default='main',
-                        help='Branch to make PR against.')
-    parser.add_argument('--timeout',
-                        type=int,
-                        default=30)
-
-
-@repo_args.apply()
-@commands.register('puppet', 'sync-ldap',
-                   help='Sync to LDAP')
-def sync_ldap(args: argparse.Namespace):
-    try:
-        _sync_ldap(args)
-    except Exception as e:
-        logger = logging.getLogger(__name__)
-        logger.critical(traceback.format_exc())
-
-
-def _sync_ldap(args: argparse.Namespace):
-    logger = logging.getLogger(__name__)
-    
-    site_data = SiteData(args.site_dir,
-                         common_root=args.global_dir,
-                         key_dir=args.key_dir,
-                         load=False)
-    lm = LDAPManager(config=args.config.ldap['ucdavis'])
-
-    lock = site_data.lock(args.timeout)
-    with lock:
-    #    gh = Gh(working_dir=site_data.common.root)
-    #    git = Git(working_dir=site_data.common.root)
-    #    git.checkout(branch=args.base_branch)()
-    #    git.clean(force=True, exclude=os.path.basename(lock.lock_file))()
-    #    git.pull()()
-#
-        site_data.load()
-
-    #   working_branch, pr_title = branch_name_title()
-    #    git.checkout(branch=working_branch, create=True)()
-
-    #    fqdn = socket.getfqdn()
-        
-    #    get_commit = git.rev_parse()
-    #    start_commit = get_commit().strip()
-
-        for user_name, user_record in site_data.iter_users():
-            synced_record = update_user_from_ldap(lm, user_name, user_record)
-            if synced_record != user_record:
-                logger.info(user_record)
-                logger.info(synced_record)
-                site_data.common.create_user(user_name, synced_record, force=True)
-
-
-def update_user_from_ldap(lm: LDAPManager, user_name: str, user_record: PuppetUserRecord) -> PuppetUserRecord:
-    logger = logging.getLogger(__name__)
-    logger.info(f'LDAP: Querying {user_name} against {lm.config.servers}')
-    status, response = lm._search_user([user_name])
-    if status is False:
-        logger.warning(f'LDAP: query failed on user {user_name}; response: {response}')
-        return user_record
-    #logger.info(f'LDAP: got status {status}, response {response}')
-    ldap_record = response[0]['attributes']
-    if 'mail' not in ldap_record:
-        logger.warning(f'LDAP: user {user_name} has no "mail" attribute.')
-        return user_record
-    if ldap_record['displayName'] != user_record.fullname or \
-        ldap_record['mail'][0] != user_record.email:
-        return PuppetUserRecord.from_other(user_record,
-                                           fullname=ldap_record['displayName'],
-                                           email=ldap_record['mail'][0])
-    else:
-        return user_record
